# gen/shared/src/main/scala/contract/model/CreateDraftResponse.scala
/** Contract Management API API for managing draft contracts
  *
  * OpenAPI spec version: 1.0.0
  *
  * Contact: team@openapitools.org
  *
  * NOTE: This class is auto generated by OpenAPI Generator.
  *
  * https://openapi-generator.tech
  */

// this model was generated using model.mustache
package contract.model
import scala.util.control.NonFatal

// see https://com-lihaoyi.github.io/upickle/
import upickle.default.{ReadWriter => RW, macroRW}
import upickle.default.*

case class CreateDraftResponse(
    /* The reference identifier for the first counterparty. */
    firstCounterpartyRef: Option[String] = None,

    /* The reference identifier for the second counterparty. */
    secondCounterpartyRef: Option[String] = None
) {

  def asJson: String = asData.asJson

  def asData: CreateDraftResponseData = {
    CreateDraftResponseData(
      firstCounterpartyRef = firstCounterpartyRef.getOrElse(""),
      secondCounterpartyRef = secondCounterpartyRef.getOrElse("")
    )
  }

}

object CreateDraftResponse {

  given RW[CreateDraftResponse] =
    CreateDraftResponseData.readWriter.bimap[CreateDraftResponse](_.asData, _.asModel)

  enum Fields(fieldName: String) extends Field(fieldName) {
    case firstCounterpartyRef  extends Fields("firstCounterpartyRef")
    case secondCounterpartyRef extends Fields("secondCounterpartyRef")
  }

}